# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Iron Temple API
  version: 0.1.0
  description: >
    REST API for the Iron Temple weight-lifting tracker. All weights are in
    pounds (lb).

    Authentication is a session cookie (`it_session`), issued by
    `POST /auth/login` and revoked by `POST /auth/logout`. Every endpoint
    requires it except `/health`, the `/auth/*` endpoints, and avatar images.
    Registration is open only while no account exists: the first sign-up claims
    the install and closes the door behind it.

    Programs, days and exercises are shared — the prescription is the same for
    everyone. Sessions are per-user: what you lifted is yours, and progression
    is computed from your own history alone.

    Programs are made of days (e.g. Workout A / B); each day prescribes exercises
    with sets, reps, and a starting weight. A session is a performed instance of
    a day, with one logged set per prescribed set.

    Progression is linear and computed server-side from session history:
    +5 lb per session on squat, bench, row, and overhead press; +10 lb on
    deadlift; a 10% deload after three consecutive failed sessions on a lift.
servers:
  - url: /api/v1

# Applies to every operation; the handful of public ones override with
# `security: []`.
security:
  - cookieAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Create the first account and sign in
      description: >
        Succeeds only while no account exists; afterwards it returns 403.
        Sessions logged before any account existed are adopted by this user.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created and signed in
          headers:
            Set-Cookie:
              description: Session cookie (`it_session`)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Registration is closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in
          headers:
            Set-Cookie:
              description: >
                Session cookie (`it_session`). Without `rememberMe` it is a
                browser-session cookie; with it, `Max-Age` is 60 days.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: Too many failed attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Sign out, revoking this session server-side
      responses:
        "204":
          description: Signed out
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me:
    get:
      tags: [users]
      operationId: getMe
      summary: The signed-in user's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [users]
      operationId: updateMe
      summary: Update display name, avatar colour, or current program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/password:
    put:
      tags: [users]
      operationId: changePassword
      summary: Change the password, revoking all other sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Not signed in, or the current password was wrong
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me/avatar:
    post:
      tags: [users]
      operationId: uploadAvatar
      summary: Upload an avatar image
      description: >
        Accepts a PNG or JPEG of at most 256 KB and 1024x1024 pixels. The image
        is decoded and re-encoded before storage, which drops EXIF metadata.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [avatar]
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Avatar"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Image too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [users]
      operationId: deleteAvatar
      summary: Remove the avatar, falling back to the initials chip
      responses:
        "204":
          description: Removed (idempotent)
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{userId}/avatar:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [users]
      operationId: getUserAvatar
      summary: An avatar image
      description: >
        Public, because this is an `<img>` source and a browser cannot attach
        credentials to one. Responds with an ETag; honours `If-None-Match`.
      security: []
      responses:
        "200":
          description: The image
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        "304":
          description: Not modified
        "404":
          $ref: "#/components/responses/NotFound"
  /exercises:
    get:
      tags: [exercises]
      operationId: listExercises
      summary: List the exercise library
      description: >
        The seeded catalogue plus the caller's own custom movements. Someone
        else's custom exercise is never returned.
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [all, performed]
            default: all
          description: >
            all returns the whole library. performed narrows it to lifts the
            caller has actually logged, which is what the Progress page wants —
            a library's worth of untrained movements reading "no sessions yet"
            is noise.
      responses:
        "200":
          description: Exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [exercises]
      operationId: createExercise
      summary: Add a custom exercise to the library
      description: >
        Creates a movement visible only to the caller. Names are compared
        case-insensitively against the shared catalogue and the caller's own
        exercises; a collision with either is a 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /exercises/{exerciseId}:
    parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: [exercises]
      operationId: deleteExercise
      summary: Delete one of your own custom exercises
      description: >
        Seeded exercises cannot be deleted by anyone. A custom one cannot be
        deleted while it has logged sets against it — that would take a lift out
        of a finished session, which is a record — or while it is on a program
        day as assistance; remove it there first.
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /exercises/{exerciseId}/history:
    parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [exercises]
      operationId: getExerciseHistory
      summary: Per-session weight history for a lift (oldest first)
      responses:
        "200":
          description: History points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseHistoryPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs:
    get:
      tags: [programs]
      operationId: listPrograms
      summary: List available programs
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgramSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
    get:
      tags: [programs]
      operationId: getProgram
      summary: Get a program with its days and prescribed exercises
      responses:
        "200":
          description: Program detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}/days/{dayId}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/DayId"
    patch:
      tags: [programs]
      operationId: updateProgramDayWeekday
      summary: Set or clear the weekday a program day is scheduled on
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgramDayRequest"
      responses:
        "204":
          description: Updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}/days/{dayId}/assistance:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/DayId"
    post:
      tags: [programs]
      operationId: addAssistance
      summary: Attach an assistance exercise to a program day
      description: >
        Adds to the caller's own overlay on a shared program day; the program's
        prescription is not modified, and nobody else's view of the day changes.
        Takes effect on the next session started from this day — a session
        already in progress keeps the sets it was created with.


        Two 409s: the exercise is already on this day as assistance, or the
        program itself already prescribes it. The second is not a nicety —
        prescribing one lift twice collides on the session's own unique
        constraint, so the workout could not be started at all.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssistanceRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDayAssistance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}/days/{dayId}/assistance/{assistanceId}:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/DayId"
      - $ref: "#/components/parameters/AssistanceId"
    patch:
      tags: [programs]
      operationId: updateAssistance
      summary: Change an assistance exercise's sets, reps or fallback weight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssistanceRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDayAssistance"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [programs]
      operationId: removeAssistance
      summary: Remove an assistance exercise from a program day
      description: >
        Removes the plan, never the performance. Sets already logged against the
        exercise stay in the sessions that recorded them and keep counting toward
        volume and records.
      responses:
        "204":
          description: Removed
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /programs/{programId}/days/{dayId}/next-session:
    parameters:
      - $ref: "#/components/parameters/ProgramId"
      - $ref: "#/components/parameters/DayId"
    get:
      tags: [programs]
      operationId: previewNextSession
      summary: Preview the next session for a program day
      description: >
        Returns the prescription for the day with weights computed from history
        via the program's progression rules. Does not create a session.
      responses:
        "200":
          description: Prescribed next session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrescribedSession"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /racked:
    get:
      tags: [racked]
      operationId: getRacked
      summary: Statistics recap for one month or year
      description: >
        The Racked recap: headline volume, hero moments and the series behind
        each chart, for the calendar month or year containing `on`. The same
        report is rendered into the monthly and yearly recap email, so the two
        can never disagree.
      parameters:
        - name: period
          in: query
          required: false
          description: Granularity of the recap. Defaults to month.
          schema:
            type: string
            enum: [month, year]
            default: month
        - name: on
          in: query
          required: false
          description: >
            Any date inside the wanted period; the server expands it to the
            whole month or year. Defaults to today, so the recap opens on the
            period in progress.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Recap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RackedReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sessions:
    get:
      tags: [sessions]
      operationId: listSessions
      summary: List performed sessions (most recent first)
      parameters:
        - name: programId
          in: query
          required: false
          description: Filter to sessions of a given program
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Paginated session history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [sessions]
      operationId: createSession
      summary: Start a session from a program day
      description: >
        Materializes the day's prescription (with progression-computed weights)
        into a new session and its logged sets.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sessions/{sessionId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    get:
      tags: [sessions]
      operationId: getSession
      summary: Get a session with its sets
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      tags: [sessions]
      operationId: updateSession
      summary: Update session metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionRequest"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [sessions]
      operationId: deleteSession
      summary: Delete a session
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sessions/{sessionId}/finish:
    parameters:
      - $ref: "#/components/parameters/SessionId"
    post:
      tags: [sessions]
      operationId: finishSession
      summary: End a session
      description: >
        Stamps finishedAt and marks the session over, whether or not every set
        hit its target. Idempotent: finishing an already-finished session leaves
        the original timestamp in place. A session also becomes over on its own
        12 hours after it was started.
      responses:
        "200":
          description: The finished session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sessions/{sessionId}/sets/{setId}:
    parameters:
      - $ref: "#/components/parameters/SessionId"
      - $ref: "#/components/parameters/SetId"
    patch:
      tags: [sessions]
      operationId: updateSessionSet
      summary: Log or amend a single set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionSetRequest"
      responses:
        "200":
          description: Updated set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionSet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: it_session
      description: >
        Opaque session token. Set by the server as an HttpOnly, SameSite=Lax
        cookie — it is never readable by, or set from, JavaScript, so clients
        neither send nor manage it explicitly.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Unauthorized:
      description: Not signed in, or the session has expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, avatarColor, isAdmin, hasAvatar]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          example: ada
        displayName:
          type: string
          example: Ada Lovelace
        avatarColor:
          type: string
          description: >
            Hex colour for the initials chip. Empty means the UI derives one
            from the id.
          example: "#b026ff"
        isAdmin:
          type: boolean
        currentProgramId:
          type: integer
          format: int64
          description: >
            The program the user last opened, which the app lands on instead of
            the picker. Absent until they open one; the UI then falls back to
            the program of their most recent session.
        hasAvatar:
          type: boolean
          description: Whether to render the image or the initials chip.
        avatarEtag:
          type: string
          description: >
            Append to the avatar URL as a cache-buster so a new upload shows
            immediately.
...
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[A-Za-z0-9._-]+$"
        displayName:
          type: string
          maxLength: 64
          description: Defaults to the username when omitted.
        password:
          type: string
          format: password
          minLength: 8
        rememberMe:
          type: boolean
          default: false
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        rememberMe:
          type: boolean
          default: false
          description: >
            Issues a 60-day cookie whose expiry slides forward as it is used,
            instead of one that dies with the browser session.
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 8
    Avatar:
      type: object
      required: [etag]
      properties:
        etag:
          type: string
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          example: session 42 not found
...